add avoir

+++ src/Manager/DemandeManager.php
@@ -4,7 +4,7 @@
  * @Author: Patrick &lt;&lt; rapaelec@gmail.com &gt;&gt; 
  * @Date: 2019-04-17 13:14:01 
  * @Last Modified by: Patrick << rapaelec@gmail.com >>
- * @Last Modified time: 2019-11-27 13:37:08
+ * @Last Modified time: 2019-12-09 10:55:08
  */
 namespace App\Manager;
 
@@ -317,6 +317,40 @@ class DemandeManager
             $snappy = new Pdf('/usr/local/bin/wkhtmltopdf');
             $filename = "Facture";
             $html = $this->twig->render("payment/facture.html.twig", $params);
+            $output = $snappy->getOutputFromHtml($html);
+            
+            $filefinal = file_put_contents($file, $output);
+        }
+        
+        return $file;
+    }
+
+    public function generateAvoir(Demande &$demande)
+    {
+        if (is_null($demande->getAvoir())){
+            Throw new Exception("La demande n'a pas encore été retracter");
+        }
+        $folder = $demande->getGeneratedCerfaPath();
+        $file = $demande->getGeneratedAvoirPathFile();
+        $params = $this->getTitulaireParams($demande);
+        $params = array_merge(['demande' => $demande], $params);
+        // create directory
+        if (!is_dir($folder)) mkdir($folder, 0777, true);
+        // end create file 
+        // save Avoir before generate number
+        if(
+            !is_null($demande->getAvoir()) &&
+            $demande->getAvoir()->getFullPath() != $file
+        ) {
+            $demande->getAvoir()->setFullPath($file);
+            $this->saveDemande($demande);
+        }
+        // end sav Avoir before generate number
+        // get facture if not exist
+        if (!is_file($file)) { // attente de finalité du process
+            $snappy = new Pdf('/usr/local/bin/wkhtmltopdf');
+            $filename = "Facture";
+            $html = $this->twig->render("avoir/avoir.pdf.twig", $params);
             $output = $snappy->getOutputFromHtml($html);
             
             $filefinal = file_put_contents($file, $output);
